\name{MOJOV.regTermTest}
\alias{MOJOV.regTermTest}
\title{
A internal function.
A test from survey package.
}
\description{
A test from survey package for linear regression.
}
\usage{
MOJOV.regTermTest(x, y, ...)
}
\arguments{
  \item{x}{
A genotype vector.
]}
  \item{y}{
A phenotype vector.
}
  \item{\dots}{
Other arguments for regTermTest from survey. See ?regTermTest}
}

\author{
Ke-Hao Wu}
