\name{MOJOV.saws}
\alias{MOJOV.saws}
\title{
A test from saws package.
}
\description{
A test from saws package for linear regression.
}
\usage{
MOJOV.saws(x, y, ...)
}
\arguments{
  \item{x}{
A genotype vector.
}
  \item{y}{
A phenotype vector.
}
  \item{\dots}{
Other arguments for regTermTest from survey. See ?saws
}
}
