\name{MOJOV.Summary-class}
\Rdversion{1.1}
\docType{class}
\alias{MOJOV.Summary-class}
\alias{show,MOJOV.Summary-method}

\title{Class \code{"MOJOV.Summary"}}
\description{
A class for MOJOV package, and you can create it by call \code{new("MOJOV.Summary",...)}
}




\author{
Ke-Hao Wu
}

\examples{
showClass("MOJOV.Summary")
}
\keyword{classes}
