\name{MOJOV.wald.test}
\alias{MOJOV.wald.test}
\title{
A test from aod package.
}
\description{
A test from aod package for linear regression.
}
\usage{
MOJOV.wald.test(x, y, ...)
}
\arguments{
  \item{x}{
A genotype vector.
}
  \item{y}{
A phenotype vector.
}
  \item{\dots}{
Other arguments for regTermTest from survey. See ?regTermTest}
}

\author{
Ke-Hao Wu
}
