\name{MOJOV.weight}
\alias{MOJOV.weight}
\title{
Providing a weight vector.
}
\description{
A function to provide weight data for analsysis.}
\usage{
MOJOV.weight(x = NULL, y = NULL, weightMethod = "ChuanhuaXin")
}
\arguments{
  \item{x}{
A genotype vector.
}
  \item{y}{
A phenotype vector.
}
  \item{weightMethod}{
Specify the weighted method for CMC, It only provide "ChuanhuaXin" method in this version.}
}

\author{
Ke-Hao Wu
}
